Se agrega una clase madre con funciones de registro para abstraer el código. Se agrega un mensaje que notifica si contraseña/usuario fueron erroneos en login.
